Simulate a dataset with true dates, observed dates and error indicators

Simulate a dataset with true dates, observed dates and error indicators

simulate_data(
  n_per_group,
  delay_map,
  delay_params,
  error_params,
  date_range,
  simul_error = FALSE
)

Arguments

n_per_group

Vector of number of individuals to simulate in each group.

delay_map

A data frame that defines the delays between events. It must contain the columns from (character), to (character), and group (list of numeric group IDs).

delay_params

A data frame containing the parameters (delay_mean, delay_cv) for each delay. Consider combining delay_map and delay_params?

error_params

A list containing prop_missing_data and prob_error.

date_range

A vector of two integer dates for the simulation range.

simul_error

Boolean. If TRUE, simulates missing and erroneous data.

Value

A list with three data frames: true_data, observed_data, and error_indicators, each including id and group columns.

Examples

# Define the delay_map data frame
delay_map <- data.frame(
  from = c("onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  group = I(list(1:4, 2, 3, 3, 4, 4))
)

# Define the delay parameters data frame
delay_params <- data.frame(
  group = c(1:4, 2, 3, 3, 4, 4),
  from = c("onset", "onset", "onset", "onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "report", "report", "report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  delay_mean = c(10, 10, 10, 10, 15, 7, 20, 7, 12),
  delay_cv = c(0.3, 0.3, 0.3, 0.3, 0.4, 0.2, 0.5, 0.2, 0.3)
)

# Define other parameters
n_per_group <- rep(10, max(delay_params$group))
error_params <- list(prop_missing_data = 0.2, prob_error = 0.05)
date_range <- as.integer(as.Date(c("2025-03-01", "2025-09-01")))

# Run simulation
set.seed(1)
sim_result <- simulate_data(
  n_per_group = n_per_group,
  delay_map = delay_map,
  delay_params = delay_params,
  error_params = error_params,
  date_range = date_range,
  simul_error = TRUE
)

sim_result$true_data
#>    id group    onset hospitalisation   report    death discharge
#> 1   1     1 20215.57              NA 20229.42       NA        NA
#> 2   2     1 20190.63              NA 20196.20       NA        NA
#> 3   3     1 20232.77              NA 20242.30       NA        NA
#> 4   4     1 20221.78              NA 20236.27       NA        NA
#> 5   5     1 20184.27              NA 20192.99       NA        NA
#> 6   6     1 20312.34              NA 20321.76       NA        NA
#> 7   7     1 20319.83              NA 20330.70       NA        NA
#> 8   8     1 20310.72              NA 20319.63       NA        NA
#> 9   9     1 20191.73              NA 20202.82       NA        NA
#> 10 10     1 20269.41              NA 20283.36       NA        NA
#> 11 11     2 20281.33              NA 20292.05 20294.82        NA
#> 12 12     2 20169.34              NA 20182.02 20180.78        NA
#> 13 13     2 20217.39              NA 20229.31 20230.26        NA
#> 14 14     2 20222.76              NA 20230.02 20232.88        NA
#> 15 15     2 20170.64              NA 20183.88 20189.65        NA
#> 16 16     2 20297.60              NA 20308.36 20308.11        NA
#> 17 17     2 20296.13              NA 20305.52 20323.68        NA
#> 18 18     2 20250.75              NA 20259.97 20260.61        NA
#> 19 19     2 20195.51              NA 20206.50 20205.52        NA
#> 20 20     2 20187.10              NA 20200.16 20201.74        NA
#> 21 21     3 20148.50        20154.15 20156.40       NA  20185.35
#> 22 22     3 20191.39        20197.84 20202.93       NA  20229.60
#> 23 23     3 20282.19        20290.05 20296.96       NA  20313.17
#> 24 24     3 20247.28        20254.46 20259.98       NA  20280.04
#> 25 25     3 20224.27        20229.92 20230.79       NA  20267.86
#> 26 26     3 20246.95        20255.13 20257.66       NA  20291.90
#> 27 27     3 20197.17        20203.40 20206.80       NA  20208.71
#> 28 28     3 20263.58        20272.43 20272.51       NA  20281.06
#> 29 29     3 20164.93        20171.69 20171.21       NA  20186.29
#> 30 30     3 20196.28        20204.29 20203.19       NA  20224.77
#> 31 31     4 20277.05        20282.48 20283.61 20290.34        NA
#> 32 32     4 20263.10        20268.42 20274.01 20279.82        NA
#> 33 33     4 20332.18        20340.33 20341.00 20355.57        NA
#> 34 34     4 20252.50        20257.57 20263.00 20267.38        NA
#> 35 35     4 20175.81        20183.80 20187.57 20193.42        NA
#> 36 36     4 20173.27        20180.02 20184.66 20186.75        NA
#> 37 37     4 20184.28        20192.64 20197.70 20200.18        NA
#> 38 38     4 20219.84        20224.81 20229.12 20242.01        NA
#> 39 39     4 20295.33        20303.81 20304.83 20316.72        NA
#> 40 40     4 20258.24        20266.30 20264.58 20274.32        NA
sim_result$observed_data
#>    id group      onset hospitalisation     report      death  discharge
#> 1   1     1 2025-05-07            <NA> 2025-05-21       <NA>       <NA>
#> 2   2     1 2025-04-12            <NA> 2025-04-18       <NA>       <NA>
#> 3   3     1 2025-05-24            <NA> 2025-06-03       <NA>       <NA>
#> 4   4     1 2025-05-13            <NA> 2025-05-28       <NA>       <NA>
#> 5   5     1 2025-04-06            <NA>       <NA>       <NA>       <NA>
#> 6   6     1 2025-08-12            <NA> 2025-08-21       <NA>       <NA>
#> 7   7     1 2025-08-19            <NA> 2025-08-30       <NA>       <NA>
#> 8   8     1 2025-08-10            <NA> 2025-08-19       <NA>       <NA>
#> 9   9     1       <NA>            <NA> 2025-04-24       <NA>       <NA>
#> 10 10     1 2025-06-30            <NA>       <NA>       <NA>       <NA>
#> 11 11     2 2025-07-12            <NA> 2025-07-23 2025-07-25       <NA>
#> 12 12     2       <NA>            <NA> 2025-04-04 2025-04-02       <NA>
#> 13 13     2 2025-05-09            <NA> 2025-05-21 2025-05-22       <NA>
#> 14 14     2       <NA>            <NA> 2025-05-22 2025-05-24       <NA>
#> 15 15     2 2025-04-25            <NA>       <NA>       <NA>       <NA>
#> 16 16     2 2025-07-28            <NA> 2025-08-08       <NA>       <NA>
#> 17 17     2 2025-07-27            <NA> 2025-08-05 2025-08-23       <NA>
#> 18 18     2 2025-06-11            <NA> 2025-06-20 2025-07-14       <NA>
#> 19 19     2 2025-04-17            <NA> 2025-04-28 2025-04-27       <NA>
#> 20 20     2 2025-04-09            <NA> 2025-04-22 2025-04-23       <NA>
#> 21 21     3 2025-03-01      2025-03-07 2025-03-09       <NA> 2025-04-25
#> 22 22     3 2025-04-13      2025-04-19       <NA>       <NA> 2025-05-21
#> 23 23     3 2025-07-13      2025-07-21 2025-07-27       <NA> 2025-08-13
#> 24 24     3 2025-06-08            <NA> 2025-06-20       <NA>       <NA>
#> 25 25     3 2025-05-16      2025-05-21 2025-05-22       <NA> 2025-06-28
#> 26 26     3 2025-06-07      2025-06-16 2025-06-18       <NA> 2025-07-22
#> 27 27     3 2025-04-19      2025-08-28 2025-04-28       <NA>       <NA>
#> 28 28     3       <NA>      2025-07-03 2025-05-10       <NA> 2025-07-12
#> 29 29     3       <NA>      2025-03-24 2025-03-24       <NA> 2025-04-08
#> 30 30     3       <NA>      2025-04-26       <NA>       <NA> 2025-05-16
#> 31 31     4 2025-07-08      2025-07-13 2025-07-14 2025-07-21       <NA>
#> 32 32     4 2025-06-24      2025-07-11       <NA> 2025-07-10       <NA>
#> 33 33     4 2025-09-01      2025-09-09 2025-09-10 2025-09-24       <NA>
#> 34 34     4 2025-06-13      2025-06-18 2025-06-23 2025-06-28       <NA>
#> 35 35     4       <NA>      2025-04-05 2025-04-09 2025-04-15       <NA>
#> 36 36     4 2025-03-26      2025-04-02 2025-04-06 2025-04-08       <NA>
#> 37 37     4 2025-04-06      2025-04-14 2025-04-19       <NA>       <NA>
#> 38 38     4       <NA>      2025-05-16 2025-05-21 2025-06-03       <NA>
#> 39 39     4 2025-07-26            <NA> 2025-08-04 2025-08-16       <NA>
#> 40 40     4 2025-06-19      2025-06-27 2025-06-25 2025-07-05       <NA>
sim_result$error_indicators # true error indicators
#>    id group onset hospitalisation report death discharge
#> 1   1     1 FALSE              NA  FALSE    NA        NA
#> 2   2     1 FALSE              NA  FALSE    NA        NA
#> 3   3     1 FALSE              NA  FALSE    NA        NA
#> 4   4     1 FALSE              NA  FALSE    NA        NA
#> 5   5     1 FALSE              NA     NA    NA        NA
#> 6   6     1 FALSE              NA  FALSE    NA        NA
#> 7   7     1 FALSE              NA  FALSE    NA        NA
#> 8   8     1 FALSE              NA  FALSE    NA        NA
#> 9   9     1    NA              NA  FALSE    NA        NA
#> 10 10     1 FALSE              NA     NA    NA        NA
#> 11 11     2 FALSE              NA  FALSE FALSE        NA
#> 12 12     2    NA              NA  FALSE FALSE        NA
#> 13 13     2 FALSE              NA  FALSE FALSE        NA
#> 14 14     2    NA              NA  FALSE FALSE        NA
#> 15 15     2  TRUE              NA     NA    NA        NA
#> 16 16     2 FALSE              NA  FALSE    NA        NA
#> 17 17     2 FALSE              NA  FALSE FALSE        NA
#> 18 18     2 FALSE              NA  FALSE  TRUE        NA
#> 19 19     2 FALSE              NA  FALSE FALSE        NA
#> 20 20     2 FALSE              NA  FALSE FALSE        NA
#> 21 21     3 FALSE           FALSE  FALSE    NA      TRUE
#> 22 22     3 FALSE           FALSE     NA    NA     FALSE
#> 23 23     3 FALSE           FALSE  FALSE    NA     FALSE
#> 24 24     3 FALSE              NA  FALSE    NA        NA
#> 25 25     3 FALSE           FALSE  FALSE    NA     FALSE
#> 26 26     3 FALSE           FALSE  FALSE    NA     FALSE
#> 27 27     3 FALSE            TRUE  FALSE    NA        NA
#> 28 28     3    NA           FALSE   TRUE    NA     FALSE
#> 29 29     3    NA           FALSE  FALSE    NA     FALSE
#> 30 30     3    NA           FALSE     NA    NA     FALSE
#> 31 31     4 FALSE           FALSE  FALSE FALSE        NA
#> 32 32     4 FALSE            TRUE     NA FALSE        NA
#> 33 33     4 FALSE           FALSE  FALSE FALSE        NA
#> 34 34     4 FALSE           FALSE  FALSE FALSE        NA
#> 35 35     4    NA           FALSE  FALSE FALSE        NA
#> 36 36     4 FALSE           FALSE  FALSE FALSE        NA
#> 37 37     4 FALSE           FALSE  FALSE    NA        NA
#> 38 38     4    NA           FALSE  FALSE FALSE        NA
#> 39 39     4 FALSE              NA  FALSE FALSE        NA
#> 40 40     4 FALSE           FALSE  FALSE FALSE        NA